Refactor project: Rename project (Quiz to QuizAPI) and app name (quiz_app to quiz)
