+++ QuizAPI/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for Quiz project.
+Django settings for QuizAPI project.
 
 Generated by 'django-admin startproject' using Django 4.2.2.
 
@@ -38,7 +38,7 @@ INSTALLED_APPS = [
     'rest_framework',
     'rest_framework.authtoken',
     'account',
-    'quiz_app'
+    'quiz'
 ]
 
 AUTH_USER_MODEL = 'account.UserProfile'
@@ -53,7 +53,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'Quiz.urls'
+ROOT_URLCONF = 'QuizAPI.urls'
 
 TEMPLATES = [
     {
@@ -72,7 +72,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'Quiz.wsgi.application'
+WSGI_APPLICATION = 'QuizAPI.wsgi.application'
 
 # Database
 # https://docs.djangoproject.com/en/4.2/ref/settings/#databases
