+++ quiz/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class QuizAppConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'quiz_app'
+    name = 'quiz'
